♻️refactor: Table View components
