+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { ContentBody, ContentLayout, Control } from "@/components";
+import { Control } from "@/components";
 import { useAtomValue, useSetAtom } from "jotai";
 import { fontFamilyAtom, loadConfigAtom } from "@/store";
 import { useEffect } from "react";
@@ -6,7 +6,7 @@ import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 import { ThemeProvider } from "@/components";
 import { cn } from "@/libs";
-import DemoPage from "@/components/payments/page";
+import TableView from "@/components/TableView/TableView";
 
 function App() {
   const loadConfig = useSetAtom(loadConfigAtom);
@@ -19,16 +19,15 @@ function App() {
   return (
     <main
       className={cn(
-        "h-full overflow-hidden antialiased select-none text-foreground rounded-lg bg-background",
+        "h-screen overflow-auto antialiased select-none text-foreground rounded-lg bg-background",
         fontFamily,
       )}
     >
       <ThemeProvider defaultTheme="system">
         <Control />
-        <ContentLayout>
-          {/* <ContentBody /> */}
-          <DemoPage />
-        </ContentLayout>
+        <div className="container mx-auto py-10">
+          <TableView />
+        </div>
         <ToastContainer
           position="bottom-right"
           autoClose={5000}
